+++ packages/enhanced-img/CHANGELOG.md
@@ -1,5 +1,53 @@
 # @sveltejs/enhanced-img
 
+## 0.10.0
+### Minor Changes
+
+
+- feat: re-export the `Picture` type from `vite-imagetools` ([#15225](https://github.com/sveltejs/kit/pull/15225))
+
+## 0.9.3
+### Patch Changes
+
+
+- fix: ensure compatibility with vite-plugin-svelte@7 (use api.filter instead of api.idFilter) ([#15179](https://github.com/sveltejs/kit/pull/15179))
+
+## 0.9.2
+### Patch Changes
+
+
+- chore: remove duplicate caching layer ([#14988](https://github.com/sveltejs/kit/pull/14988))
+
+## 0.9.1
+### Patch Changes
+
+
+- fix: update vite-imagetools for caching fixes to avoid crashes ([#14976](https://github.com/sveltejs/kit/pull/14976))
+
+## 0.9.0
+### Minor Changes
+
+
+- chore(deps): update dependency vite-imagetools to v9 ([#14608](https://github.com/sveltejs/kit/pull/14608))
+
+## 0.8.5
+### Patch Changes
+
+
+- fix: warn rather than crash when non-enhanced image dynamically passed to `enhanced:img` ([#14845](https://github.com/sveltejs/kit/pull/14845))
+
+## 0.8.4
+### Patch Changes
+
+
+- fix: add script block only when there are imports to add ([#14604](https://github.com/sveltejs/kit/pull/14604))
+
+## 0.8.3
+### Patch Changes
+
+
+- chore: update "homepage" field in package.json ([#14579](https://github.com/sveltejs/kit/pull/14579))
+
 ## 0.8.2
 ### Patch Changes